+++ Runtime/IRCMessages.cs
@@ -33,6 +33,34 @@ namespace Incredulous.Twitch
 
         public ChatterBadge[] badges = new ChatterBadge[0];
         public List<ChatterEmote> emotes = new List<ChatterEmote>();
+
+        /// <summary>
+        /// Returns whether the tags contain a given emote.
+        /// </summary>
+        public bool ContainsEmote(string emote)
+        {
+            foreach (ChatterEmote e in emotes)
+            {
+                if (e.id == emote)
+                    return true;
+            }
+
+            return false;
+        }
+
+        /// <summary>
+        /// Returns whether the tags contain a given badge.
+        /// </summary>
+        public bool HasBadge(string badge)
+        {
+            foreach (ChatterBadge b in badges)
+            {
+                if (b.id == badge)
+                    return true;
+            }
+
+            return false;
+        }
     }
 
 }