+++ Runtime/TwitchConnection.SendThread.cs
@@ -1,3 +1,5 @@
+using System;
+using System.Collections.Generic;
 using System.Collections.Concurrent;
 using System.Threading;
 using UnityEngine;
@@ -10,38 +12,42 @@ namespace Incredulous.Twitch
         private ConcurrentQueue<string> priorityOutputQueue = new ConcurrentQueue<string>();
         private ConcurrentQueue<string> outputQueue = new ConcurrentQueue<string>();
 
-        private const int twitchRateLimitSleepTime = 1750;
-
         /// <summary>
         /// The IRC output process which will run on the send thread.
         /// </summary>
         private void SendProcess()
         {
             var stream = tcpClient.GetStream();
+            RateLimit rateLimit;
 
             //Read loop
             while (continueThreads)
             {
-                int sleepTime = writeInterval;
+                // Send prioritized outputs
+                while (priorityOutputQueue.TryDequeue(out var priorityOutput))
+                    stream.WriteLine(priorityOutput, debugIRC);
 
-                if (!priorityOutputQueue.IsEmpty)
+                // Get the current rate limit
+                lock (rateLimitLock)
+                    rateLimit = chatRateLimit;
+
+                // Clear timestamps that are older than the rate limit period
+                var minTime = DateTime.Now - rateLimit.timeSpan;
+                while (outputTimestamps.TryPeek(out var next) && next < minTime)
+                    outputTimestamps.TryDequeue(out _);
+
+                // Send outputs up to the rate limit
+                while (!outputQueue.IsEmpty && outputTimestamps.Count < rateLimit.count)
                 {
-                    // Send all outputs from priorityOutputQueue
-                    while (priorityOutputQueue.TryDequeue(out var output))
-                        stream.WriteLine(output, debugIRC);
-                }
-                else if (!outputQueue.IsEmpty)
-                {
-                    // Send next output from outputQueue
                     if (outputQueue.TryDequeue(out var output))
                     {
                         stream.WriteLine(output, debugIRC);
-                        sleepTime = twitchRateLimitSleepTime;
+                        outputTimestamps.Enqueue(DateTime.Now);
                     }
                 }
 
                 // Sleep for a short while before checking again
-                Thread.Sleep(sleepTime);
+                Thread.Sleep(writeInterval);
             }
 
             if (debugThreads)
@@ -51,13 +57,23 @@ namespace Incredulous.Twitch
         /// <summary>
         /// Sends a ping to the server.
         /// </summary>
-        public void Ping() => SendCommand("PING :tmi.twitch.tv", true);
+        public void Ping() => SendCommand("PING :tmi.twitch.tv");
 
         /// <summary>
-        /// Queues a command to be sent to the IRC server. All prioritzed commands will be sent before non-prioritized commands.
+        /// Queues a command to be sent to the IRC server. Prioritzed commands will be sent without regard for rate limits.
         /// </summary>
         public void SendCommand(string command, bool prioritized = false)
         {
+            // For non-prioritized outputs, send a warning if the output will surpass the rate limit
+            if (!prioritized)
+            {
+                lock (rateLimitLock)
+                {
+                    if (outputTimestamps.Count + 1 > chatRateLimit.count)
+                        alertQueue.Enqueue(ConnectionAlert.RateLimitWarning);
+                }
+            }
+
             // Place command in respective queue
             if (prioritized)
                 priorityOutputQueue.Enqueue(command);
@@ -73,6 +89,13 @@ namespace Incredulous.Twitch
             // Message can't be empty
             if (message.Length <= 0)
                 return;
+
+            // Send a warning if the output will surpass the rate limit
+            lock (rateLimitLock)
+            {
+                if (outputTimestamps.Count + 1 > chatRateLimit.count)
+                    alertQueue.Enqueue(ConnectionAlert.RateLimitWarning);
+            }
 
             // Place message in queue
             outputQueue.Enqueue("PRIVMSG #" + twitchCredentials.channel + " :" + message);